Constraints

Constraint Templates

Constraint templates help simplify data wrangling across multiple Power Flow formulations by providing an abstraction layer between the network data and network constraint definitions. The constraint template's job is to extract the required parameters from a given network data structure and pass the data as named arguments to the Power Flow formulations.

These templates should be defined over GenericPowerModel and should not refer to model variables. For more details, see the files: core/constraint_template.jl and core/constraint.jl.

Generator Constraints

PowerModels.constraint_active_gen_setpoint — Function.

source

source

PowerModels.constraint_reactive_gen_setpoint — Function.

source

source

do nothing, this model does not have reactive variables

source

Bus Constraints

Setpoint Constraints

PowerModels.constraint_theta_ref — Function.

source

source

source

Do nothing, no way to represent this in these variables

source

source

Do nothing, no way to represent this in these variables

source

PowerModels.constraint_voltage_magnitude_setpoint — Function.

source

source

do nothing, this model does not have voltage variables

source

KCL Constraints

PowerModels.constraint_kcl_shunt — Function.

source

source

source

source

source

source

PowerModels.constraint_kcl_shunt_ne — Function.

source

source

source

source

source

Branch Constraints

Ohm's Law Constraints

PowerModels.constraint_ohms_yt_from — Function.

source

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

source

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

source

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

source

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

source

PowerModels.constraint_ohms_yt_to — Function.

source

Do nothing, this model is symmetric

source

source

source

PowerModels.constraint_ohms_y_from — Function.

source

Creates Ohms constraints for AC models (y post fix indicates that Y values are in rectangular form)

source

PowerModels.constraint_ohms_y_to — Function.

source

source

On/Off Ohm's Law Constraints

PowerModels.constraint_ohms_yt_from_on_off — Function.

source

source

source

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

source

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

source

PowerModels.constraint_ohms_yt_to_on_off — Function.

source

source

Do nothing, this model is symetric

source

source

source

PowerModels.constraint_ohms_yt_from_ne — Function.

source

source

source

source

PowerModels.constraint_ohms_yt_to_ne — Function.

source

source

Do nothing, this model is symmetric

source

source

source

Current

PowerModels.constraint_power_magnitude_sqr — Function.

source

source

PowerModels.constraint_power_magnitude_link — Function.

source

source

Thermal Limit Constraints

PowerModels.constraint_thermal_limit_from — Function.

source

source

source

source

PowerModels.constraint_thermal_limit_to — Function.

source

source

source

Do nothing, this model is symmetric

source

PowerModels.constraint_thermal_limit_from_on_off — Function.

source

source

Generic on/off thermal limit constraint

source

PowerModels.constraint_thermal_limit_to_on_off — Function.

source

source

nothing to do, from handles both sides

source

source

PowerModels.constraint_thermal_limit_from_ne — Function.

source

source

source

PowerModels.constraint_thermal_limit_to_ne — Function.

source

source

nothing to do, from handles both sides

source

source

Phase Angle Difference Constraints

PowerModels.constraint_phase_angle_difference — Function.

source

source

source

source

source

source

PowerModels.constraint_phase_angle_difference_on_off — Function.

source

source

source

source

PowerModels.constraint_phase_angle_difference_ne — Function.

source

source

source

source

Loss Constraints

PowerModels.constraint_loss_lb — Function.

source

source

Commonly Used Constraints

The following methods generally assume that the model contains p and q values for branches line flows and bus flow conservation.

Generic thermal limit constraint

constraint_thermal_limit_from(pm::GenericPowerModel, f_idx, rate_a)
constraint_thermal_limit_to(pm::GenericPowerModel, t_idx, rate_a)

Generic on/off thermal limit constraint

constraint_thermal_limit_from_on_off(pm::GenericPowerModel, i, f_idx, rate_a)
constraint_thermal_limit_to_on_off(pm::GenericPowerModel, i, t_idx, rate_a)
constraint_thermal_limit_from_ne(pm::GenericPowerModel, i, f_idx, rate_a)
constraint_thermal_limit_to_ne(pm::GenericPowerModel, i, t_idx, rate_a)
constraint_active_gen_setpoint(pm::GenericPowerModel, i, pg)
constraint_reactive_gen_setpoint(pm::GenericPowerModel, i, qg)