Power Model

All methods for constructing powermodels should be defined on the following type:

PowerModels.AbstractPowerModel — Type

root of the power formulation type hierarchy

which utilizes the following (internal) functions:

PowerModels.build_ref — Function

used for building ref without the need to build a initialize an AbstractPowerModel

PowerModels.calc_buspair_parameters — Function

compute bus pair level data, can be run on data or ref data structures