Objective

PowerModels.check_cost_models — Method

Checks that all cost models are of the same type

PowerModels.check_dcline_cost_models — Method

Checks that all dcline cost models are of the same type

PowerModels.check_gen_cost_models — Method

Checks that all generator cost models are of the same type

PowerModels.objective_min_fuel_and_flow_cost — Method

PowerModels.objective_min_fuel_and_flow_cost_polynomial — Method

PowerModels.objective_min_fuel_and_flow_cost_pwl — Method

PowerModels.objective_min_fuel_cost — Method

PowerModels.objective_min_fuel_cost_polynomial — Method

PowerModels.objective_min_fuel_cost_pwl — Method

PowerModels.objective_variable_dc_cost — Function

adds pdccost variables and constraints

PowerModels.objective_variable_pg_cost — Function

adds pg_cost variables and constraints

PowerModels._objective_min_fuel_and_flow_cost_polynomial_linquad — Function

Adds lifted variables to turn a quadatic objective into a linear one; needed for conic solvers that only support linear objectives

PowerModels._objective_min_fuel_and_flow_cost_polynomial_linquad — Method

PowerModels._objective_min_fuel_and_flow_cost_polynomial_nl — Method

PowerModels._objective_min_fuel_cost_polynomial_linquad — Method

PowerModels._objective_min_fuel_cost_polynomial_nl — Method