File IO

File IO

General Data Formats

PowerModels.parse_file — Function.

source

PowerModels.parse_json — Function.

source

Matpower Data Files

The following two methods are the main exported methods for parsing matpower data files:

parse_matpower
parse_matpower_data

We also provide the following (internal) helper methods:

standardize_cost_order
update_branch_transformer_settings
merge_generator_cost_data
merge_bus_name_data
parse_cell
parse_matrix
parse_matlab_data
split_line
add_line_delimiter
extract_assignment
extract_mpc_assignment
type_value
type_array
build_typed_dict
extend_case_data
mp_data_to_pm_data
split_loads_shunts

PTI Data Files (PSS/E)

Note: This feature is currently in development, and only the parsing of PTI files into a Dict containing all of the raw data directly from the PTI file is currently supported. No data conversion to a PowerModels supported format is available yet.

The following method is the main exported method for parsing PTI data files:

PowerModels.parse_pti — Function.

parse_pti(filename)

Open PTI raw file given by filename, passing the file contents as a string to the main PTI parser, returning a Dict of all the data parsed into the proper types.

source

The following internal helper methods are also provided:

PowerModels.get_pti_sections — Function.

get_pti_sections()

Returns Array of the names of the sections, in the order that they appear in a PTI file, v33+

source

PowerModels.get_pti_dtypes — Function.

get_pti_dtypes(field_name)

Returns OrderedDict of data types for PTI file section given by field_name, as enumerated by PSS/E Program Operation Manual

source

PowerModels.parse_line_element! — Function.

parse_line_element!(data, elements, section)

Parses a single "line" of data elements from a PTI file, as given by elements which is an array of the line, typically split at ,. Elements are parsed into data types given by section and saved into data::Dict

source

PowerModels.add_section_data! — Function.

add_section_data!(pti_data, section_data, section)

Adds section_data::Dict, which contains all parsed elements of a PTI file section given by section, into the parent pti_data::Dict

source

PowerModels.get_line_elements — Function.

get_line_elements(line)

Uses regular expressions to extract all separate data elements from a line of a PTI file and populate them into an Array{String}. Comments, typically indicated at the end of a line with a '/' character, are also extracted separately, and Array{Array{String}, String} is returned.

source

PowerModels.parse_pti_data — Function.

parse_pti_data(data_string, sections)

Parse a PTI raw file into a Dict, given the data_string of the file and a list of the sections in the PTI file (typically given by default by get_pti_sections().

source