Relaxation Schemes

PowerModels.cut_complex_product_and_angle_difference — Method

A valid inequality for the product of two complex variables with magnitude and angle difference bounds.

In the literature this constraints are called the Lifted Nonlinear Cuts (LNCs).

@misc{1512.04644, Author = {Carleton Coffrin and Hassan Hijazi and Pascal Van Hentenryck}, Title = {Strengthening the SDP Relaxation of AC Power Flows with Convex Envelopes, Bound Tightening, and Lifted Nonlinear Cuts}, Year = {2015}, Eprint = {arXiv:1512.04644}, }

PowerModels.cut_product_replicates — Method

A valid inequality for when the same product of two variables occurs in two different higher order products (e.g. trilinear terms).

@misc{1809.04565, Author = {Kaarthik Sundar and Harsha Nagarajan and Sidhant Misra and Mowen Lu and Carleton Coffrin and Russell Bent}, Title = {Optimization-Based Bound Tightening using a Strengthened QC-Relaxation of the Optimal Power Flow Problem}, Year = {2018}, Eprint = {arXiv:1809.04565}, }

PowerModels.relaxation_cos — Method

general relaxation of a cosine term, in -pi/2 to pi/2

PowerModels.relaxation_cos_on_off — Method

general relaxation of a cosine term, in -pi/2 to pi/2

PowerModels.relaxation_sin — Method

general relaxation of a sine term, in -pi/2 to pi/2

PowerModels.relaxation_sin_on_off — Method

general relaxation of a sine term, in -pi/2 to pi/2