Variables

Variables

We provide the following methods to provide a compositional approach for defining common variables used in power flow models. These methods should always be defined over "GenericPowerModel".

PowerModels.getstart — Function.

extracts the start value

source

PowerModels.variable_active_generation — Method.

variable: pg[j] for j in gen

source

PowerModels.variable_active_line_flow — Method.

variable: p[l,i,j] for (l,i,j) in arcs

source

PowerModels.variable_active_line_flow_ne — Method.

variable: -ne_branch[l]["rate_a"] <= p_ne[l,i,j] <= ne_branch[l]["rate_a"] for (l,i,j) in ne_arcs

source

PowerModels.variable_generation — Method.

generates variables for both active and reactive generation

source

PowerModels.variable_line_flow — Method.

source

PowerModels.variable_line_flow_ne — Method.

generates variables for both active and reactive line_flow_ne

source

PowerModels.variable_line_indicator — Method.

variable: 0 <= line_z[l] <= 1 for l in branches

source

PowerModels.variable_line_ne — Method.

variable: 0 <= line_ne[l] <= 1 for l in branches

source

PowerModels.variable_phase_angle — Method.

variable: t[i] for i in buses

source

PowerModels.variable_reactive_generation — Method.

variable: qq[j] for j in gen

source

PowerModels.variable_reactive_line_flow — Method.

variable: q[l,i,j] for (l,i,j) in arcs

source

PowerModels.variable_reactive_line_flow_ne — Method.

variable: -ne_branch[l]["rate_a"] <= q_ne[l,i,j] <= ne_branch[l]["rate_a"] for (l,i,j) in ne_arcs

source

PowerModels.variable_voltage_magnitude — Method.

variable: v[i] for i in buses

source

PowerModels.variable_voltage_magnitude_sqr — Method.

variable: w[i] >= 0 for i in buses

source

PowerModels.variable_voltage_magnitude_sqr_from_on_off — Method.

variable: 0 <= w_from[l] <= buses[branches[l]["f_bus"]]["vmax"]^2 for l in branches

source

PowerModels.variable_voltage_magnitude_sqr_to_on_off — Method.

variable: 0 <= w_to[l] <= buses[branches[l]["t_bus"]]["vmax"]^2 for l in branches

source