Problem Specifications

This section provides links to the formulation agnostic problem specifications.

Power Flow (PF)

For additional discussion on the power flow problem please see, Power Flow Computations.

Bus Injection Models (BIM)

PowerModels.build_pf — Function

specification of the formulation agnostic Power Flow model

Branch Flow Models (BFM)

PowerModels.build_pf_bf — Function

Current Flow Models (CFM)

PowerModels.build_pf_iv — Function

Optimal Power Flow (OPF)

Bus Injection Models (BIM)

PowerModels.build_opf — Function

PowerModels.build_mn_opf — Function

PowerModels.build_mn_opf_strg — Function

PowerModels.build_opf_ptdf — Function

Branch Flow Models (BFM)

PowerModels.build_opf_bf — Function

PowerModels.build_mn_opf_bf_strg — Function

Build multinetwork branch flow storage OPF

Current Flow Models (CFM)

PowerModels.build_opf_iv — Function

Optimal Power Balance (OPB)

A copper-plate approximation of the OPF problem

PowerModels.build_opb — Function

Optimal Transmission Switching (OTS)

PowerModels.build_ots — Function

Note

if the branch status is 0 in the input, it is out of service and forced to 0 in OTS

Transmission Network Expansion Planning (TNEP)

PowerModels.build_tnep — Function

the general form of the tnep optimization model