Formulation Details

Formulation Details

This section provides references to understand the formulations as provided by PowerModels. The list is not meant as a literature discussion, but to give the main starting points to understand the implementation of the formulations.

Molzahn, D., & Hiskens, I. (forthcoming). A Survey of Relaxations and Approximations of the Power Flow Equations. Foundations and Trends in Electric Energy Systems

Coffrin, C., & Roald, L. (2018). Convex relaxations in power system optimization: a brief introduction. [Math.OC], 1–5. Retrieved from http://arxiv.org/abs/1807.07227

Coffrin, C., Hijazi, H., & Van Hentenryck, P. (2016). The QC relaxation: a theoretical and computational study on optimal power flow. IEEE Trans. Power Syst., 31(4), 3008–3018. https://doi.org/10.1109/TPWRS.2015.2463111

Notes on the mathematical model for all formulations

PowerModels implements a slightly generalized version of the AC Optimal Power Flow problem from Matpower, as discussed in The PowerModels Mathematical Model and presented here.

In the next subsections the differences between PowerModels' bus and branch models and those commonly used in the literature are discussed. Consideration is given to these differences when implementing formulations from articles.

Standardized branch model

The branch model is standardized as follows:

An idealized (lossless) transformer at the from side of the branch (immediately on node $i$) with a fixed, complex-value voltage transformation (i.e. tap and shift)

Followed by a pi-section with complex-valued line shunt admittance, where the from and to side shunt can have different values

A branch is uniquely defined by a tuple $(l,i,j)$ where $l$ is the line index, $i$ is the from node, and $j$ is the to node.

Thermal limits are defined in apparent power, and are defined at both ends of a line

Each branch has a phase angle difference constraint

Nevertheless, in the literature, the following can be observed:

The to and from side line shunts are equal

The line shunt admittance is a pure susceptance (equivalent to shunt conductance set to 0)

A branch in a grid without parallel lines is uniquely defined by a tuple $(i,j)$ where $i$ is the from node, and $j$ is the to node.

Thermal limits are defined in current (total or series), complex power limits are approximated as a regular polygon, ...

Thermal limits are defined only at the from (or to) side

Phase angle difference constraints are not included

Furthermore, "lifted nonlinear cuts" are used to improve the accuracy of PAD constraints for all formulations in the lifted S-W variable space:

Coffrin, C., Hijazi, H., & Van Hentenryck, P. (2017). Strengthening the SDP relaxation of ac power flows with convex envelopes, bound tightening, and valid inequalities. IEEE Trans. Power Syst., 32(5), 3549–3558. https://doi.org/10.1109/TPWRS.2016.2634586

Standardized bus model

The bus model is standardized as follows:

A bus defines a complex power balance for all the sets lines, generators, loads, bus shunts connected to it, i.e. one can define multiple load and shunt components on each bus $i$

Nevertheless, in the literature, a simplified bus model is often used:

Only a single (aggregated) load per bus is supported

Only a single (aggregated) bus shunt per bus is supported

Exact Non-Convex Models

PowerModels.ACPPowerModel — Type.

AC power flow formulation with polar bus voltage variables.

The seminal reference of AC OPF:

@article{carpentier1962contribution,
  title={Contribution to the economic dispatch problem},
  author={Carpentier, J},
  journal={Bulletin de la Societe Francoise des Electriciens},
  volume={3},
  number={8},
  pages={431--447},
  year={1962}
}

History and discussion:

@techreport{Cain2012,
  author = {Cain, Mary B and {O' Neill}, Richard P and Castillo, Anya},
  title = {{History of optimal power flow and formulations}},
  year = {2012}
  pages = {1--36},
  url = {https://www.ferc.gov/industries/electric/indus-act/market-planning/opf-papers/acopf-1-history-formulation-testing.pdf}
}

source

PowerModels.ACRPowerModel — Type.

AC power flow formulation with rectangular bus voltage variables.

@techreport{Cain2012,
  author = {Cain, Mary B and {O' Neill}, Richard P and Castillo, Anya},
  pages = {1--36},
  title = {{History of optimal power flow and formulations}},
  url = {https://www.ferc.gov/industries/electric/indus-act/market-planning/opf-papers/acopf-1-history-formulation-testing.pdf}
  year = {2012}
}

source

PowerModels.ACTPowerModel — Type.

AC power flow formulation (nonconvex) with variables for voltage angle, voltage magnitude squared, and real and imaginary part of voltage crossproducts. A tangens constraint is added to represent meshed networks in an exact manner.

@ARTICLE{4349090,
  author={R. A. Jabr},
  title={A Conic Quadratic Format for the Load Flow Equations of Meshed Networks},
  journal={IEEE Transactions on Power Systems},
  year={2007},
  month={Nov},
  volume={22},
  number={4},
  pages={2285-2286},
  doi={10.1109/TPWRS.2007.907590},
  ISSN={0885-8950}
}

source

Linear Approximations

PowerModels.DCPPowerModel — Type.

Linearized 'DC' power flow formulation with polar voltage variables.

@ARTICLE{4956966,
  author={B. Stott and J. Jardim and O. Alsac},
  journal={IEEE Transactions on Power Systems},
  title={DC Power Flow Revisited},
  year={2009},
  month={Aug},
  volume={24},
  number={3},
  pages={1290-1300},
  doi={10.1109/TPWRS.2009.2021235},
  ISSN={0885-8950}
}

source

PowerModels.NFAPowerModel — Type.

The an active power only network flow approximation, also known as the transportation model.

source

Quadratic Approximation

PowerModels.DCPLLPowerModel — Type.

source

PowerModels.LPACCPowerModel — Type.

The LPAC Cold-Start AC Power Flow Approximation.

Note that the LPAC Cold-Start model requires the least amount of information but is also the least accurate variant of the LPAC formulations. If a nominal AC operating point is available, the LPAC Warm-Start model will provide improved accuracy.

The original publication suggests to use polyhedral outer approximations for the cosine and line thermal lit constraints. Given the recent improvements in MIQCQP solvers, this implementation uses quadratic functions for those constraints.

@article{doi:10.1287/ijoc.2014.0594,
  author = {Coffrin, Carleton and Van Hentenryck, Pascal},
  title = {A Linear-Programming Approximation of AC Power Flows},
  journal = {INFORMS Journal on Computing},
  volume = {26},
  number = {4},
  pages = {718-734},
  year = {2014},
  doi = {10.1287/ijoc.2014.0594},
  eprint = {https://doi.org/10.1287/ijoc.2014.0594}
}

source

Quadratic Relaxations

PowerModels.SOCWRPowerModel — Type.

Second-order cone relaxation of bus injection model of AC OPF.

The implementation casts this as a convex quadratically constrained problem.

@article{1664986,
  author={R. A. Jabr},
  title={Radial distribution load flow using conic programming},
  journal={IEEE Transactions on Power Systems},
  year={2006},
  month={Aug},
  volume={21},
  number={3},
  pages={1458-1459},
  doi={10.1109/TPWRS.2006.879234},
  ISSN={0885-8950}
}

source

PowerModels.SOCWRConicPowerModel — Type.

Second-order cone relaxation of bus injection model of AC OPF.

This implementation casts the problem as a convex conic problem.

source

PowerModels.QCWRPowerModel — Type.

"Quadratic-Convex" relaxation of AC OPF

@Article{Hijazi2017,
  author="Hijazi, Hassan and Coffrin, Carleton and Hentenryck, Pascal Van",
  title="Convex quadratic relaxations for mixed-integer nonlinear programs in power systems",
  journal="Mathematical Programming Computation",
  year="2017",
  month="Sep",
  volume="9",
  number="3",
  pages="321--367",
  issn="1867-2957",
  doi="10.1007/s12532-016-0112-z",
  url="https://doi.org/10.1007/s12532-016-0112-z"
}

source

PowerModels.QCWRTriPowerModel — Type.

"Quadratic-Convex" relaxation of AC OPF with convex hull of triple product

@Article{Hijazi2017,
  author="Hijazi, Hassan and Coffrin, Carleton and Hentenryck, Pascal Van",
  title="Convex quadratic relaxations for mixed-integer nonlinear programs in power systems",
  journal="Mathematical Programming Computation",
  year="2017",
  month="Sep",
  volume="9",
  number="3",
  pages="321--367",
  issn="1867-2957",
  doi="10.1007/s12532-016-0112-z",
  url="https://doi.org/10.1007/s12532-016-0112-z"
}

source

PowerModels.SOCBFPowerModel — Type.

Second-order cone relaxation of branch flow model

The implementation casts this as a convex quadratically constrained problem.

@INPROCEEDINGS{6425870,
  author={M. Farivar and S. H. Low},
  title={Branch flow model: Relaxations and convexification},
  booktitle={2012 IEEE 51st IEEE Conference on Decision and Control (CDC)},
  year={2012},
  month={Dec},
  pages={3672-3679},
  doi={10.1109/CDC.2012.6425870},
  ISSN={0191-2216}
}

Extended as discussed in:

@misc{1506.04773,
  author = {Carleton Coffrin and Hassan L. Hijazi and Pascal Van Hentenryck},
  title = {DistFlow Extensions for AC Transmission Systems},
  year = {2018},
  eprint = {arXiv:1506.04773},
  url = {https://arxiv.org/abs/1506.04773}
}

source

PowerModels.SOCBFConicPowerModel — Type.

source

SDP Relaxation

PowerModels.SDPWRMPowerModel — Type.

Semi-definite relaxation of AC OPF

Originally proposed by:

@article{BAI2008383,
  author = "Xiaoqing Bai and Hua Wei and Katsuki Fujisawa and Yong Wang",
  title = "Semidefinite programming for optimal power flow problems",
  journal = "International Journal of Electrical Power & Energy Systems",
  volume = "30",
  number = "6",
  pages = "383 - 392",
  year = "2008",
  issn = "0142-0615",
  doi = "https://doi.org/10.1016/j.ijepes.2007.12.003",
  url = "http://www.sciencedirect.com/science/article/pii/S0142061507001378",
}

First paper to use "W" variables in the BIM of AC OPF:

@INPROCEEDINGS{6345272,
  author={S. Sojoudi and J. Lavaei},
  title={Physics of power networks makes hard optimization problems easy to solve},
  booktitle={2012 IEEE Power and Energy Society General Meeting},
  year={2012},
  month={July},
  pages={1-8},
  doi={10.1109/PESGM.2012.6345272},
  ISSN={1932-5517}
}

source

PowerModels.SparseSDPWRMPowerModel — Type.

Sparsity-exploiting semidefinite relaxation of AC OPF

Proposed in:

@article{doi:10.1137/S1052623400366218,
  author = {Fukuda, M. and Kojima, M. and Murota, K. and Nakata, K.},
  title = {Exploiting Sparsity in Semidefinite Programming via Matrix Completion I: General Framework},
  journal = {SIAM Journal on Optimization},
  volume = {11},
  number = {3},
  pages = {647-674},
  year = {2001},
  doi = {10.1137/S1052623400366218},
  URL = {https://doi.org/10.1137/S1052623400366218},
  eprint = {https://doi.org/10.1137/S1052623400366218}
}

Original application to OPF by:

@ARTICLE{6064917,
  author={R. A. Jabr},
  title={Exploiting Sparsity in SDP Relaxations of the OPF Problem},
  journal={IEEE Transactions on Power Systems},
  volume={27},
  number={2},
  pages={1138-1139},
  year={2012},
  month={May},
  doi={10.1109/TPWRS.2011.2170772},
  ISSN={0885-8950}
}

source