Power Model

All methods for constructing powermodels should be defined on the following type:

PowerModels.AbstractPowerModel — Type

root of the power formulation type hierarchy

which utilizes the following (internal) functions:

PowerModels.build_ref — Function

used for building ref without the need to build a initialize an AbstractPowerModel

Missing docstring.

Missing docstring for buspair_parameters!. Check Documenter's build log for details.