Objective

Objective

PowerModels.check_cost_models — Method.

Checks that all cost models are of the same type

source

PowerModels.check_dcline_cost_models — Method.

Checks that all dcline cost models are of the same type

source

PowerModels.check_gen_cost_models — Method.

Checks that all generator cost models are of the same type

source

PowerModels.objective_min_fuel_and_flow_cost — Method.

source

PowerModels.objective_min_fuel_and_flow_cost_polynomial — Method.

source

PowerModels.objective_min_fuel_and_flow_cost_pwl — Method.

source

PowerModels.objective_min_fuel_cost — Method.

source

PowerModels.objective_min_fuel_cost_polynomial — Method.

source

PowerModels.objective_min_fuel_cost_pwl — Method.

source

PowerModels.objective_variable_dc_cost — Method.

adds pdccost variables and constraints

source

PowerModels.objective_variable_pg_cost — Method.

adds pg_cost variables and constraints

source

PowerModels._objective_min_fuel_and_flow_cost_polynomial_linquad — Method.

source

PowerModels._objective_min_fuel_and_flow_cost_polynomial_linquad — Method.

Adds lifted variables to turn a quadatic objective into a linear one; needed for conic solvers that only support linear objectives

source

PowerModels._objective_min_fuel_and_flow_cost_polynomial_nl — Method.

source

PowerModels._objective_min_fuel_cost_polynomial_linquad — Method.

source

PowerModels._objective_min_fuel_cost_polynomial_nl — Method.

source