Objective

Objective

PowerModels._objective_min_gen_polynomial_fuel_cost_linquad — Method.

source

PowerModels._objective_min_gen_polynomial_fuel_cost_nl — Method.

source

PowerModels._objective_min_polynomial_fuel_cost_linquad — Method.

source

PowerModels._objective_min_polynomial_fuel_cost_linquad — Method.

Adds lifted variables to turn a quadatic objective into a linear one; needed for conic solvers that only support linear objectives

source

PowerModels._objective_min_polynomial_fuel_cost_nl — Method.

source

PowerModels.check_cost_models — Method.

Checks that all cost models are present and of the same type

source

PowerModels.get_lines — Method.

compute lines in m and b from from pwl cost models data is a list of components

source

PowerModels.objective_min_fuel_cost — Method.

source

PowerModels.objective_min_gen_fuel_cost — Method.

source

PowerModels.objective_min_gen_polynomial_fuel_cost — Method.

source

PowerModels.objective_min_gen_pwl_fuel_cost — Method.

source

PowerModels.objective_min_polynomial_fuel_cost — Method.

source

PowerModels.objective_min_pwl_fuel_cost — Method.

source

PowerModels.objective_tnep_cost — Method.

Cost of building branches

source

PowerModels.slope_intercepts — Method.

compute m and b from points pwl points

source