Objective

Objective

PowerModels.check_cost_models — Method.

Checks that all cost models are present and of the same type

source

PowerModels.check_polynomial_cost_models — Method.

Checks that all cost models are polynomials, quadratic or less

source

PowerModels.get_lines — Method.

compute lines in m and b from from pwl cost models data is a list of components

source

PowerModels.objective_min_fuel_cost — Function.

source

PowerModels.objective_min_polynomial_fuel_cost — Function.

source

PowerModels.objective_min_polynomial_fuel_cost — Method.

source

PowerModels.objective_min_pwl_fuel_cost — Function.

source

PowerModels.objective_tnep_cost — Function.

Cost of building branches

source

PowerModels.slope_intercepts — Method.

compute m and b from points pwl points

source