File IO

File IO

General Data Formats

PowerModels.parse_file — Function.

parse_file(file; import_all)

Parses a Matpower .m file or PTI (PSS(R)E-v33) .raw file into a PowerModels data structure. All fields from PTI files will be imported if import_all is true (Default: false).

source

Parses the iostream from a file

source

PowerModels.parse_json — Function.

Parses json from iostream or string

source

Matpower Data Files

The following method is the main exported methods for parsing Matpower data files:

PowerModels.parse_matpower — Function.

Parses the matpwer data from either a filename or an IO object

source

We also provide the following (internal) helper methods:

PowerModels.export_matpower — Method.

Export power network data in the matpower format

source

PowerModels.export_matpower — Method.

Export power network data in the matpower format

source

PowerModels.parse_matpower — Method.

Parses the matpwer data from either a filename or an IO object

source

PowerModels._add_dcline_costs! — Method.

adds dcline costs, if gen costs exist

source

PowerModels._check_keys — Method.

source

PowerModels._export_cost_data — Method.

Export cost data

source

PowerModels._export_extra_data — Function.

Export fields of a component type

source

PowerModels._get_default — Function.

Get a default value for dict entry

source

PowerModels._matpower_to_powermodels! — Method.

Converts a Matpower dict into a PowerModels dict

source

PowerModels._merge_bus_name_data! — Method.

merges bus name data into buses, if names exist

source

PowerModels._merge_generator_cost_data! — Method.

merges generator cost functions into generator data, if costs exist

source

PowerModels._merge_generic_data! — Method.

merges Matpower tables based on the table extension syntax

source

PowerModels._mp2pm_branch! — Method.

sets all branch transformer taps to 1.0, to simplify branch models

source

PowerModels._mp2pm_dcline! — Method.

adds pmin and pmax values at to and from buses

source

PowerModels._mp_cost_data — Method.

source

PowerModels._parse_matpower_string — Method.

source

PowerModels._split_loads_shunts! — Method.

_split_loads_shunts!(data)

Seperates Loads and Shunts in data under separate "load" and "shunt" keys in the PowerModels data format. Includes references to originating bus via "loadbus" and "shuntbus" keys, respectively.

source

PTI Data Files (PSS/E)

Note: This feature supports the parsing and conversion of PTI files into a PowerModels format for the following power network components: buses, loads, shunts (fixed and approximation of switched), branches, two-winding and three-winding transformers (incl. magnetizing admittance), generators, two-terminal dc lines, and voltage source converter HVDC lines.

The following method is the main exported method for parsing PSS(R)E v33 specified PTI data files:

PowerModels.parse_pti — Function.

parse_pti(filename::String)

Open PTI raw file given by filename, returning a Dict of the data parsed into the proper types.

source

parse_pti(io::IO)

Reads PTI data in io::IO, returning a Dict of the data parsed into the proper types.

source

PowerModels.parse_psse — Function.

Parses directly from file

source

Parses directly from iostream

source

The following internal helper methods are also provided:

PowerModels.parse_psse — Method.

Parses directly from iostream

source

PowerModels.parse_psse — Method.

Parses directly from file

source

PowerModels._create_starbus_from_transformer — Method.

create_starbus(pm_data, transformer)

Creates a starbus from a given three-winding transformer. "sourceid" is given by `["busi", "name", "I", "J", "K", "CKT"]` where "bus_i" and "name" are the modified names for the starbus, and "I", "J", "K" and "CKT" come from the originating transformer, in the PSS(R)E transformer specification.

source

PowerModels._find_max_bus_id — Method.

_find_max_bus_id(pm_data)

Returns the maximum bus id in pm_data

source

PowerModels._get_bus_value — Method.

_get_bus_value(bus_i, field, pm_data)

Returns the value of field of bus_i from the PowerModels data. Requires "bus" Dict to already be populated.

source

PowerModels._import_remaining! — Method.

Imports remaining keys from data_in into data_out, excluding keys in exclude

source

PowerModels._init_bus! — Method.

_init_bus!(bus, id)

Initializes a bus of id id with default values given in the PSS(R)E specification.

source

PowerModels._psse2pm_branch! — Method.

_psse2pm_branch!(pm_data, pti_data)

Parses PSS(R)E-style Branch data into a PowerModels-style Dict. "source_id" is given by ["I", "J", "CKT"] in PSS(R)E Branch specification.

source

PowerModels._psse2pm_bus! — Method.

_psse2pm_bus!(pm_data, pti_data)

Parses PSS(R)E-style Bus data into a PowerModels-style Dict. "source_id" is given by ["I", "NAME"] in PSS(R)E Bus specification.

source

PowerModels._psse2pm_dcline! — Method.

_psse2pm_dcline!(pm_data, pti_data)

Parses PSS(R)E-style Two-Terminal and VSC DC Lines data into a PowerModels compatible Dict structure by first converting them to a simple DC Line Model. For Two-Terminal DC lines, "sourceid" is given by ["IPR", "IPI", "NAME"] in the PSS(R)E Two-Terminal DC specification. For Voltage Source Converters, "sourceid" is given by ["IBUS1", "IBUS2", "NAME"], where "IBUS1" is "IBUS" of the first converter bus, and "IBUS2" is the "IBUS" of the second converter bus, in the PSS(R)E Voltage Source Converter specification.

source

PowerModels._psse2pm_generator! — Method.

_psse2pm_generator!(pm_data, pti_data)

Parses PSS(R)E-style Generator data in a PowerModels-style Dict. "source_id" is given by ["I", "ID"] in PSS(R)E Generator specification.

source

PowerModels._psse2pm_load! — Method.

_psse2pm_load!(pm_data, pti_data)

Parses PSS(R)E-style Load data into a PowerModels-style Dict. "source_id" is given by ["I", "ID"] in the PSS(R)E Load specification.

source

PowerModels._psse2pm_shunt! — Method.

_psse2pm_shunt!(pm_data, pti_data)

Parses PSS(R)E-style Fixed and Switched Shunt data into a PowerModels-style Dict. "source_id" is given by ["I", "ID"] for Fixed Shunts, and ["I", "SWREM"] for Switched Shunts, as given by the PSS(R)E Fixed and Switched Shunts specifications.

source

PowerModels._psse2pm_transformer! — Method.

_psse2pm_transformer!(pm_data, pti_data)

Parses PSS(R)E-style Transformer data into a PowerModels-style Dict. "source_id" is given by ["I", "J", "K", "CKT", "winding"], where "winding" is 0 if transformer is two-winding, and 1, 2, or 3 for three-winding, and the remaining keys are defined in the PSS(R)E Transformer specification.

source

PowerModels._pti_to_powermodels! — Method.

_pti_to_powermodels!(pti_data)

Converts PSS(R)E-style data parsed from a PTI raw file, passed by pti_data into a format suitable for use internally in PowerModels. Imports all remaining data from the PTI file if import_all is true (Default: false).

source

PowerModels.parse_pti — Method.

parse_pti(io::IO)

Reads PTI data in io::IO, returning a Dict of the data parsed into the proper types.

source

PowerModels.parse_pti — Method.

parse_pti(filename::String)

Open PTI raw file given by filename, returning a Dict of the data parsed into the proper types.

source

PowerModels._add_section_data! — Method.

add_section_data!(pti_data, section_data, section)

Internal function. Adds section_data::Dict, which contains all parsed elements of a PTI file section given by section, into the parent pti_data::Dict

source

PowerModels._get_component_property — Method.

_get_component_property(section, ret, search_field, search_value)

Internal function. Finds a component in section where search_field == search_value and returns ret from that component.

source

PowerModels._get_line_elements — Method.

_get_line_elements(line)

Internal function. Uses regular expressions to extract all separate data elements from a line of a PTI file and populate them into an Array{String}. Comments, typically indicated at the end of a line with a '/' character, are also extracted separately, and Array{Array{String}, String} is returned.

source

PowerModels._get_pti_default — Method.

_get_pti_default(section, field, data, component)

Internal function. Returns a default value in section for field in component from data.

source

PowerModels._get_pti_dtypes — Method.

_get_pti_dtypes(field_name)

Internal function. Returns array of data types for PTI file section given by field_name, as enumerated by PSS/E Program Operation Manual.

source

PowerModels._get_pti_sections — Method.

_get_pti_sections()

Internal function. Returns Array of the names of the sections, in the order that they appear in a PTI file, v33

source

PowerModels._parse_line_element! — Method.

_parse_line_element!(data, elements, section)

Internal function. Parses a single "line" of data elements from a PTI file, as given by elements which is an array of the line, typically split at ,. Elements are parsed into data types given by section and saved into data::Dict.

source

PowerModels._parse_pti_data — Method.

_parse_pti_data(data_string, sections)

Internal function. Parse a PTI raw file into a Dict, given the data_string of the file and a list of the sections in the PTI file (typically given by default by get_pti_sections().

source

PowerModels._populate_defaults! — Method.

_populate_defaults!(pti_data)

Internal function. Populates empty fields with PSS(R)E PTI v33 default values

source