File IO

File IO

General Data Formats

PowerModels.parse_file — Function.

source

PowerModels.parse_json — Function.

source

Matpower Data Files

The following two methods are the main exported methods for parsing matpower data files:

PowerModels.parse_matpower — Function.

source

PowerModels.parse_matpower_data — Function.

source

We also provide the following (internal) helper methods:

PowerModels.standardize_cost_order — Function.

ensures all costs functions are quadratic and reverses their order

source

PowerModels.update_branch_transformer_settings — Function.

sets all line transformer taps to 1.0, to simplify line models

source

PowerModels.merge_generator_cost_data — Function.

merges generator cost functions into generator data, if costs exist

source

PowerModels.merge_bus_name_data — Function.

merges bus name data into buses, if names exist

source

PowerModels.parse_cell — Function.

source

PowerModels.parse_matrix — Function.

source

PowerModels.parse_matlab_data — Function.

source

PowerModels.split_line — Function.

source

PowerModels.add_line_delimiter — Function.

source

PowerModels.extract_assignment — Function.

source

PowerModels.extract_mpc_assignment — Function.

source

PowerModels.type_value — Function.

Attempts to determine the type of a string extracted from a matlab file

source

PowerModels.type_array — Function.

Attempts to determine the type of an array of strings extracted from a matlab file

source

PowerModels.build_typed_dict — Function.

takes a list of list of strings and turns it into a list of typed dictionaries

source

PowerModels.extend_case_data — Function.

extends a give case data with typed dictionary data

source

PowerModels.mp_data_to_pm_data — Function.

converts arrays of objects into a dicts with lookup by "index"

source