Variables

We provide the following methods to provide a compositional approach for defining common variables used in power flow models. These methods should always be defined over "AbstractPowerModel".

PowerModels.sol_component_value_buspair — Method

map sparse buspair variables into components

PowerModels.variable_active_branch_flow — Method

variable: p[l,i,j] for (l,i,j) in arcs

PowerModels.variable_active_branch_flow_ne — Method

variable: -ne_branch[l]["rate_a"] <= p_ne[l,i,j] <= ne_branch[l]["rate_a"] for (l,i,j) in ne_arcs

PowerModels.variable_active_dcline_flow — Method

variable: p_dc[l,i,j] for (l,i,j) in arcs_dc

PowerModels.variable_active_generation — Method

variable: pg[j] for j in gen

PowerModels.variable_active_storage — Method

PowerModels.variable_active_switch_flow — Method

variable: pws[l,i,j] for (l,i,j) in arcs_sw

PowerModels.variable_branch_current_imaginary — Method

variable: ci[l,i,j] for (l,i,j) in arcs

PowerModels.variable_branch_current_real — Method

variable: cr[l,i,j] for (l,i,j) in arcs

PowerModels.variable_branch_flow — Method

PowerModels.variable_branch_flow_ne — Method

generates variables for both active and reactive branch_flow_ne

PowerModels.variable_branch_indicator — Method

variable: 0 <= z_branch[l] <= 1 for l in branches

PowerModels.variable_branch_ne — Method

variable: 0 <= branch_ne[l] <= 1 for l in branches

PowerModels.variable_branch_series_current_imaginary — Method

variable: csi[l,i,j] for (l,i,j) in arcs_from

PowerModels.variable_branch_series_current_real — Method

variable: csr[l,i,j] for (l,i,j) in arcs_from

PowerModels.variable_cosine — Method

PowerModels.variable_current_storage — Method

do nothing by default but some formulations require this

PowerModels.variable_dcline_current_imaginary — Method

variable: cidc[j] for j in dcline

PowerModels.variable_dcline_current_real — Method

variable: crdc[j] for j in dcline

PowerModels.variable_demand_factor — Method

PowerModels.variable_gen_current_imaginary — Method

variable: cig[j] for j in gen

PowerModels.variable_gen_current_real — Method

variable: crg[j] for j in gen

PowerModels.variable_generation — Method

generates variables for both active and reactive generation

PowerModels.variable_reactive_branch_flow — Method

variable: q[l,i,j] for (l,i,j) in arcs

PowerModels.variable_reactive_branch_flow_ne — Method

variable: -ne_branch[l]["rate_a"] <= q_ne[l,i,j] <= ne_branch[l]["rate_a"] for (l,i,j) in ne_arcs

PowerModels.variable_reactive_dcline_flow — Method

variable: q_dc[l,i,j] for (l,i,j) in arcs_dc

PowerModels.variable_reactive_generation — Method

variable: qq[j] for j in gen

PowerModels.variable_reactive_storage — Method

PowerModels.variable_reactive_switch_flow — Method

variable: pws[l,i,j] for (l,i,j) in arcs_sw

PowerModels.variable_shunt_factor — Method

PowerModels.variable_sine — Method

PowerModels.variable_storage — Method

variables for modeling storage units, includes grid injection and internal variables

PowerModels.variable_storage_charge — Method

PowerModels.variable_storage_complementary_indicator — Method

PowerModels.variable_storage_discharge — Method

PowerModels.variable_storage_energy — Method

PowerModels.variable_storage_mi — Method

variables for modeling storage units, includes grid injection and internal variables, with mixed int variables for charge/discharge

PowerModels.variable_switch_flow — Method

PowerModels.variable_transformer_ta — Method

variable: ta[l] for l in branches

PowerModels.variable_transformer_tm — Method

variable: 0 <= tm[l] for l in branches

PowerModels.variable_voltage_angle — Method

variable: t[i] for i in buses

PowerModels.variable_voltage_imaginary — Method

real part of the voltage variable i in buses

PowerModels.variable_voltage_magnitude — Method

variable: v[i] for i in buses

PowerModels.variable_voltage_magnitude_from_on_off — Method

variable: 0 <= vm_fr[l] <= buses[branches[l]["f_bus"]]["vmax"] for l in branches

PowerModels.variable_voltage_magnitude_sqr — Method

variable: w[i] >= 0 for i in buses

PowerModels.variable_voltage_magnitude_sqr_from_on_off — Method

variable: 0 <= w_fr[l] <= buses[branches[l]["f_bus"]]["vmax"]^2 for l in branches

PowerModels.variable_voltage_magnitude_sqr_to_on_off — Method

variable: 0 <= w_to[l] <= buses[branches[l]["t_bus"]]["vmax"]^2 for l in branches

PowerModels.variable_voltage_magnitude_to_on_off — Method

variable: 0 <= vm_to[l] <= buses[branches[l]["t_bus"]]["vmax"] for l in branches

PowerModels.variable_voltage_product — Method

PowerModels.variable_voltage_product_on_off — Method

PowerModels.variable_voltage_real — Method

real part of the voltage variable i in buses