Variables

Variables

We provide the following methods to provide a compositional approach for defining common variables used in power flow models. These methods should always be defined over "GenericPowerModel".

PowerModels.variable_active_branch_flow — Method.

variable: p[l,i,j] for (l,i,j) in arcs

source

PowerModels.variable_active_branch_flow_ne — Method.

variable: -ne_branch[l]["rate_a"] <= p_ne[l,i,j] <= ne_branch[l]["rate_a"] for (l,i,j) in ne_arcs

source

PowerModels.variable_active_dcline_flow — Method.

variable: p_dc[l,i,j] for (l,i,j) in arcs_dc

source

PowerModels.variable_active_generation — Method.

variable: pg[j] for j in gen

source

PowerModels.variable_active_storage — Method.

source

PowerModels.variable_branch_flow — Method.

source

PowerModels.variable_branch_flow_ne — Method.

generates variables for both active and reactive branch_flow_ne

source

PowerModels.variable_branch_indicator — Method.

variable: 0 <= branch_z[l] <= 1 for l in branches

source

PowerModels.variable_branch_ne — Method.

variable: 0 <= branch_ne[l] <= 1 for l in branches

source

PowerModels.variable_cosine — Method.

source

PowerModels.variable_generation — Method.

generates variables for both active and reactive generation

source

PowerModels.variable_reactive_branch_flow — Method.

variable: q[l,i,j] for (l,i,j) in arcs

source

PowerModels.variable_reactive_branch_flow_ne — Method.

variable: -ne_branch[l]["rate_a"] <= q_ne[l,i,j] <= ne_branch[l]["rate_a"] for (l,i,j) in ne_arcs

source

PowerModels.variable_reactive_dcline_flow — Method.

variable: q_dc[l,i,j] for (l,i,j) in arcs_dc

source

PowerModels.variable_reactive_generation — Method.

variable: qq[j] for j in gen

source

PowerModels.variable_reactive_storage — Method.

source

PowerModels.variable_sine — Method.

source

PowerModels.variable_storage — Method.

variables for modeling storage units, includes grid injection and internal variables

source

PowerModels.variable_storage_charge — Method.

source

PowerModels.variable_storage_discharge — Method.

source

PowerModels.variable_storage_energy — Method.

source

PowerModels.variable_voltage_angle — Method.

variable: t[i] for i in buses

source

PowerModels.variable_voltage_imaginary — Method.

real part of the voltage variable i in buses

source

PowerModels.variable_voltage_magnitude — Method.

variable: v[i] for i in buses

source

PowerModels.variable_voltage_magnitude_from_on_off — Method.

variable: 0 <= vm_fr[l] <= buses[branches[l]["f_bus"]]["vmax"] for l in branches

source

PowerModels.variable_voltage_magnitude_sqr — Method.

variable: w[i] >= 0 for i in buses

source

PowerModels.variable_voltage_magnitude_sqr_from_on_off — Method.

variable: 0 <= w_fr[l] <= buses[branches[l]["f_bus"]]["vmax"]^2 for l in branches

source

PowerModels.variable_voltage_magnitude_sqr_to_on_off — Method.

variable: 0 <= w_to[l] <= buses[branches[l]["t_bus"]]["vmax"]^2 for l in branches

source

PowerModels.variable_voltage_magnitude_to_on_off — Method.

variable: 0 <= vm_to[l] <= buses[branches[l]["t_bus"]]["vmax"] for l in branches

source

PowerModels.variable_voltage_product — Method.

source

PowerModels.variable_voltage_product_on_off — Method.

source

PowerModels.variable_voltage_real — Method.

real part of the voltage variable i in buses

source